Add restful user API
